UPDATE - Auto refresh when open the project file
