+++ src/rendererArea/homescreenitems/visibilityOptions.tsx
@@ -1,15 +1,10 @@
-import * as THREE from 'three';
 import { ChangeEvent, useEffect, useState } from "react"
 import { FoldableControlHor } from "../components/foldableControl/foldableControlHor"
 import { SceneController } from "../api/three/SceneController";
 import { useVisibilityOptionStore } from "./visibilityOptionsStore";
 import { ModelType } from "@/mainArea/models/modelType";
 import { ButtonPositive } from "../components/buttons/buttonPositive";
-import { Boring } from "@/mainArea/models/serviceModels/boring/boring";
-import { ThreeBoringPost } from "../api/three/predefinedCreations/boringPost";
-import { LayerColorConfig } from "@/mainArea/models/uimodels/layerColorConfig";
 import { useSidebarStore } from '../sidebar/sidebarStore';
-import { BoringDTO } from '@/dto/serviceModel/BoringDTO';
 
 export const VisibilityOptions = () => {
     const [isResetBoringEnabled, setRestBoringState] = useState<boolean>(false);
@@ -86,32 +81,8 @@ export const VisibilityOptions = () => {
     }
 
     const onClickResetBorings = async () => {
-        const boringFetch = await window.electronBoringDataAPI.fetchAllBorings();
-        const layerFetch = await window.electronBoringDataAPI.getAllLayerColors();
-        if(boringFetch && boringFetch.result && layerFetch && layerFetch.result && navigationIndex != 0) {
-            const threeObjects: THREE.Object3D[] = [];
-            const threeOldObjUUIDs: string[] = [];
-            const updatedDtos: BoringDTO[] = [];
-            for (const dto of boringFetch.fetchedBorings) {
-                const boring = Boring.deserialize(dto);
-                const boringPost = new ThreeBoringPost();
-                await boringPost.init();
-                
-                const layerConfig = new LayerColorConfig(layerFetch.layerColors);
-                const threeObj = await boringPost.createPostFromModel(boring, layerConfig);
-                threeOldObjUUIDs.push(dto.threeObjId);
-                threeObjects.push(threeObj);
-
-                boring.setThreeObjId(threeObj.uuid);
-                updatedDtos.push(boring.serialize());
-            }
-
-            const updateJob = await window.electronBoringDataAPI.updateBorings(updatedDtos);
-            if(updateJob && updateJob.result) {
-                SceneController.getInstance().removeObjectByUUIDs(threeOldObjUUIDs);
-                SceneController.getInstance().addObjects(threeObjects);
-                SceneController.getInstance().getViewportControl().resetCamera();
-            }           
+        if(navigationIndex != 0) {
+            SceneController.getInstance().getDataMangeService().refreshBoringPosts();
         }
     }
